fully operational

+++ ft_striteri.c
@@ -6,15 +6,15 @@
 /*   By: tblochet <tblochet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/09 11:01:51 by tblochet          #+#    #+#             */
-/*   Updated: 2024/11/10 23:06:27 by tblochet         ###   ########.fr       */
+/*   Updated: 2024/11/12 14:11:58 by tblochet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <libft.h>
 
-void ft_striteri(char *s, void (*f)(size_t, char *))
+void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 {
-	size_t i;
+	unsigned int	i;
 
 	i = -1;
 	while (s[++i])
