+++ ft_strmapi.c
@@ -6,16 +6,16 @@
 /*   By: tblochet <tblochet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/09 11:01:51 by tblochet          #+#    #+#             */
-/*   Updated: 2024/11/10 23:27:18 by tblochet         ###   ########.fr       */
+/*   Updated: 2024/11/12 14:11:45 by tblochet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <libft.h>
 
-char	*ft_strmapi(char const *s, char (*f)(size_t, char))
+char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	char	*d;
-	size_t	i;
+	char			*d;
+	unsigned int	i;
 
 	d = ft_strdup(s);
 	i = -1;
